extracted TabPanel Component
