+++ src/containers/DataPage/TabPanel.js
@@ -1,0 +1,26 @@
+import React from "react";
+import PropTypes from 'prop-types';
+import Typography from "@material-ui/core/Typography";
+
+function TabPanel(props) {
+  const { children, value, index, ...other } = props;
+
+  return (
+    <Typography
+      component="div"
+      role="tabpanel"
+      hidden={value !== index}
+      {...other}
+    >
+      {value === index && <div>{children}</div>}
+    </Typography>
+  );
+}
+
+TabPanel.propTypes = {
+  children: PropTypes.node,
+  index: PropTypes.any.isRequired,
+  value: PropTypes.any.isRequired,
+};
+
+export default TabPanel;
