+++ src/containers/DataPage/index.js
@@ -1,11 +1,9 @@
 import React, {useContext, useEffect, useState} from "react";
-import PropTypes from 'prop-types';
 
 import MapIcon from '@material-ui/icons/Map';
 import TocIcon from '@material-ui/icons/Toc';
 import Tab from "@material-ui/core/Tab";
 import Tabs from "@material-ui/core/Tabs";
-import Typography from "@material-ui/core/Typography";
 import { useQuery } from "urql";
 
 import DataTable from "../../components/DataTable";
@@ -18,27 +16,7 @@ import {useAuth0} from "../../auth/react-auth0-spa";
 import {RoleContext} from "../App";
 import {ROLES} from "../../config";
 import searchQueryDataDisplayAdapter from "../../data/searchQueryDataDisplayAdapter";
-
-function TabPanel(props) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <Typography
-      component="div"
-      role="tabpanel"
-      hidden={value !== index}
-      {...other}
-    >
-      {value === index && <div>{children}</div>}
-    </Typography>
-  );
-}
-
-TabPanel.propTypes = {
-  children: PropTypes.node,
-  index: PropTypes.any.isRequired,
-  value: PropTypes.any.isRequired,
-};
+import TabPanel from './TabPanel';
 
 const DataPage = () => {
   const { isAuthenticated } = useAuth0();
@@ -70,17 +48,17 @@ const DataPage = () => {
   return (
     <div className="data-page">
       <SearchBar
-          coords={searchCoords}
-          setCoords={setSearchCoords}
-          distance={searchDistance}
-          setDistance={setSearchDistance}
-        />
+        coords={searchCoords}
+        setCoords={setSearchCoords}
+        distance={searchDistance}
+        setDistance={setSearchDistance}
+      />
 
       <Tabs
         value={tabIdx}
         indicatorColor="primary"
         textColor="primary"
-        onChange={(e, value) => setTabIdx(value)}
+        onChange={(_, value) => setTabIdx(value)}
         aria-label="Search results view tabs"
         centered
       >
